+++ NEWS.md
@@ -5,7 +5,7 @@
 * Remove `details` argument in `script_funs`, add `name` argument.
 * `AET05` preprocessing now filters on `"(AE|CQ|SMQ)TTE"` rather than `"AETTE"`.
 * Rename the dataset `ADAETTE` in `syn_data` object to `ADSAFTTE`.
-* Use upcase variable names in `AET05` and `AET05_ALL`.
+* Use uppercase variable names in `AET05` and `AET05_ALL`.
 
 # chevron 0.2.4